UI auto populate
These code changes should speed up future development. The UI can now be auto populated. That means that when a new item is added to the list of items, an interactive UI element is created without altering index.html. The same is true for hosts, boosts and the soon to come upgrades.